fix: corrige estrutura do AuthContext separando contexto, provider e hook
