+++ praticas/aluno-online-react/src/contexts/AuthContext.jsx
@@ -1,24 +1,3 @@
-import { createContext, useState } from 'react';
+import { createContext } from 'react';
 
 export const AuthContext = createContext(null);
-
-export function AuthProvider({ children }) {
-  const [autenticado, setAutenticado] = useState(false);
-  const [usuario, setUsuario] = useState(null);
-
-  function login(dadosUsuario) {
-    setUsuario(dadosUsuario);
-    setAutenticado(true);
-  }
-
-  function logout() {
-    setUsuario(null);
-    setAutenticado(false);
-  }
-
-  return (
-    <AuthContext.Provider value={{ autenticado, usuario, login, logout }}>
-      {children}
-    </AuthContext.Provider>
-  );
-}
